requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
openai==1.40.2
python-dotenv==1.0.1
pdfplumber==0.11.4
pdfminer.six==20240706
charset-normalizer>=3.3.2